Create WSUSReset.bat
This batch script for remote admins. if deployed first will check to see if the file has been run allready using a marker file on the C drive called WSUSReset.txt
If it doesnt exist it will clear out all cache and all residule WSUS files as well as reset winsock back to default , and any Proxy's that may be on the machine.
It will also log its progress in the marker file too.